Update availability-group-az-commandline-configure-multi-subnet.md
Clarification that Windows hostname must match the name of the SQL VM resource. If this is not the case then the deployment will fail with the error:

"code": "Ext_NodeUnreachable","message": "Error: Error code: NodeUnreachable, Message: An exception occurred during a Ping request. An exception occurred during a Ping request. No such host is known."


Documenting this has been discussed on the "PG Sync - SQLVM/Arc" chat and also previously on (a very old) IcM 217887772