Denote that OPENJSON does not support json data type (yet)
The documentation incorrectly states that all JSON functions support the `json` data type in the **Function support** section, however, the **Limitations** section denotes that `OPENJSON` does *not*, which is true. Use of the `json` data type with `OPENJSON` returns an error:
> JSON data type cannot be used in OpenJson function.

This commit denotes that `OPENJSON` doesn't support the `json` data type in the **Function support** section. It should be updated when it does.