Update from-transact-sql.md to re-add incorrectly-removed OUTER keyword
Two versions of the same query are present to demonstrate that the `OUTER` keyword is optional in the `FULL OUTER JOIN` syntax. However, currently, the two versions of the same query are actually the identical query.

This change makes the two examples different again to be consistent with the point being demonstrated as inferred from the context.